% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sampling.R
\name{cluster_sampling}
\alias{cluster_sampling}
\title{Cluster Sampling}
\usage{
cluster_sampling(data, cluster_var, n_clusters)
}
\arguments{
\item{data}{A data frame representing the sampling frame.}

\item{cluster_var}{A string specifying the column name in `data` used to define clusters.}

\item{n_clusters}{Integer. The number of clusters to select randomly.}
}
\value{
A data frame containing all elements from the selected clusters.
}
\description{
This function performs cluster sampling by selecting a specified number of clusters randomly
from a given dataset and returning all elements within the selected clusters.
}
