% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_population_parameters.R
\name{estimate_population_parameters}
\alias{estimate_population_parameters}
\title{Estimate Population Parameters}
\usage{
estimate_population_parameters(sample, population_size)
}
\arguments{
\item{sample}{A numeric vector representing the sample data (subset of the population).}

\item{population_size}{Integer. The total size of the population.}
}
\value{
A data frame containing the estimated population parameters: mean,
variance, standard deviation, median, mode, and total population value.
}
\description{
This function estimates population parameters such as mean, variance,
standard deviation, median, mode, and total population value based on
a given sample and population size.
}
