% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_of_error.R
\name{margin_of_error}
\alias{margin_of_error}
\title{Margin of Error}
\usage{
margin_of_error(standard_deviation, sample_size, confidence_level)
}
\arguments{
\item{standard_deviation}{A numeric value representing the standard deviation of the population or sample.}

\item{sample_size}{A numeric value representing the sample size.}

\item{confidence_level}{A numeric value between 0 and 1 representing the confidence level (e.g., 0.95 for 95\% confidence).}
}
\value{
A numeric value representing the margin of error.
}
\description{
This function calculates the margin of error, which is an estimate of the maximum error in sample estimation.
It is useful for determining how much a sample estimate can deviate from the actual population value,
based on a given confidence level.
}
