% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{random_sampling}
\alias{random_sampling}
\title{Random Sampling}
\usage{
random_sampling(sample_frame, sample_size)
}
\arguments{
\item{sample_frame}{A data frame or vector representing the sampling frame.}

\item{sample_size}{Integer. The desired size of the sample.}
}
\value{
A data frame or vector containing the random sample.
}
\description{
This function performs random sampling from a given sample frame.
}
