% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_sampling.R
\name{stratified_sampling}
\alias{stratified_sampling}
\title{Stratified Sampling}
\usage{
stratified_sampling(data, strata_var, sample_size_per_strata)
}
\arguments{
\item{data}{A data frame representing the sampling frame.}

\item{strata_var}{A string specifying the column name in `data` used to define strata.}

\item{sample_size_per_strata}{A named vector where names are the strata levels and values are the desired sample sizes for each stratum.}
}
\value{
A data frame containing the stratified random sample.
}
\description{
This function performs stratified random sampling, where the population is divided into strata
(subgroups) based on a specified variable, and samples are taken from each stratum according to
the specified sample size per stratum.
}
