% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_sampling.R
\name{systematic_sampling}
\alias{systematic_sampling}
\title{Systematic Sampling}
\usage{
systematic_sampling(population, sample_size)
}
\arguments{
\item{population}{A vector representing the population elements.}

\item{sample_size}{Integer. The desired number of elements to sample.}
}
\value{
A vector containing the systematically selected sample.
}
\description{
This function performs systematic sampling, selecting elements from a population
at fixed intervals starting from a random point.
}
