% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sample_frame.R
\name{validate_sample_frame}
\alias{validate_sample_frame}
\title{Validate Sample Frame}
\usage{
validate_sample_frame(sample_frame)
}
\arguments{
\item{sample_frame}{A data frame representing the sample frame to validate.}
}
\value{
Messages, warnings, or errors indicating the quality of the sample frame.
}
\description{
This function checks if a given sample frame is complete and free from major biases.
It performs a series of validations, including checking for missing values, duplicate entries,
empty rows or columns, and potential bias in categorical columns.
}
\examples{
# Example of a sample frame
sample_frame <- data.frame(
  ID = 1:10,
  Attribute1 = c("A", "B", "A", "A", "B", "B", "A", "A", "B", "B"),
  Attribute2 = c(NA, 2, 3, 4, 5, 6, 7, 8, 9, 10)
)

# Validate the sample frame
validate_sample_frame(sample_frame)
}
