% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_sampling.R
\name{visualize_sampling}
\alias{visualize_sampling}
\title{Visualize Sampling}
\usage{
visualize_sampling(sample_data, sampling_method, strata = NULL, cluster = NULL)
}
\arguments{
\item{sample_data}{A data frame containing the sample data.}

\item{sampling_method}{A string indicating the sampling method. Options
are "Stratified", "Cluster", or others.}

\item{strata}{Optional. A column in the data representing strata for stratified sampling.}

\item{cluster}{Optional. A column in the data representing cluster for cluster sampling.}
}
\value{
A ggplot object representing the visualization of the sampling process.
}
\description{
This function creates visualizations to represent how samples are selected
based on the sampling method. It supports stratified sampling, cluster
sampling, or general visualizations for other sampling methods.
}
